Reverting from an invented thing
There is a "payload" field in the GitHub tables which contains this
kind of information. I though there was one like that for every event,
only every one has its own kind of information, which is not
structured and thus goes to the "payload" field which we have to
extract.